Move identity commands from vm module to role module.
