+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_validators.py
@@ -6,6 +6,9 @@
 import uuid
 from knack.util import CLIError
 from ._client_factory import _graph_client_factory
+from azure.cli.core.commands.validators import (
+    get_default_location_from_resource_group, validate_tags)
+
 
 VARIANT_GROUP_ID_ARGS = ['object_id', 'group_id', 'group_object_id']
 
@@ -80,3 +83,8 @@ def validate_cert(namespace):
         if not x509:
             raise CLIError('usage error: --cert STRING | --cert NAME --keyvault VAULT')
         namespace.cert = x509
+
+
+def process_msi_namespace(cmd, namespace):
+    get_default_location_from_resource_group(cmd, namespace)
+    validate_tags(namespace)
